--- airflow/.env
@@ -1,1 +1,0 @@
-AIRFLOW_UID=50000
