+++ scripts/ingestion/.dlt/config.toml
@@ -1,0 +1,9 @@
+# put your configuration values here
+
+[runtime]
+log_level="WARNING"  # the system log level of dlt
+# use the dlthub_telemetry setting to enable/disable anonymous usage data reporting, see https://dlthub.com/docs/reference/telemetry
+dlthub_telemetry = true
+
+[sources.filesystem]
+bucket_url = "<configure me>" # fill this in!
