+++ scripts/ingestion/dlt_agro/bin/markdown-it
@@ -1,0 +1,10 @@
+#!/bin/sh
+'''exec' "/mnt/c/Users/Dell/Data Engineering/personal project/Dlt/dlt_agro/bin/python3" "$0" "$@"
+' '''
+# -*- coding: utf-8 -*-
+import re
+import sys
+from markdown_it.cli.parse import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
