+++ scripts/ingestion/requirements.txt
@@ -1,0 +1,1 @@
+dlt[filesystem,snowflake]>=1.11.0
