+++ scripts/ingestion/snowflake/create_tables.sql
