+++ scripts/ingestion/snowflake/database_schema.sql
@@ -1,0 +1,3 @@
+CREATE DATABASE IF NOT EXISTS AGROLIFT;
+
+CREATE SCHEMA IF NOT EXISTS AGROLIFT_SCHEMA;
