+++ scripts/requirements_script.txt
@@ -1,3 +1,2 @@
-faker
 awswrangler
-pandas
+faker
